+++ src/mako/validator/rules/file/Mimetype.php
@@ -12,9 +12,6 @@ use mako\validator\rules\RuleInterface;
 use mako\validator\rules\traits\WithParametersTrait;
 use mako\validator\rules\WithParametersInterface;
 
-use function finfo_close;
-use function finfo_file;
-use function finfo_open;
 use function implode;
 use function in_array;
 use function sprintf;
@@ -42,13 +39,7 @@ class Mimetype extends Rule implements RuleInterface, WithParametersInterface
 	{
 		$mimetypes = (array) $this->getParameter('mimetype');
 
-		$info = finfo_open(FILEINFO_MIME_TYPE);
-
-		$mimetype = finfo_file($info, $value->getPathname());
-
-		finfo_close($info);
-
-		return in_array($mimetype, $mimetypes);
+		return in_array($value->getMimeType(), $mimetypes);
 	}
 
 	/**